+++ changelog.md
@@ -1,3 +1,8 @@
+# 0.8.0
+
+* Added a upload screen mostly as a fallback for when drag and drop doesn't work
+* Free-playing now supports the mouse
+
 # 0.7.1
 
 * Maybe fixed conflicts with GUI mods